convert to oracle

+++ src/main/java/myself/programing/coding/controllers/javaCode/JavaCompileController.java
@@ -185,7 +185,7 @@ public class JavaCompileController {
     }
 
     @GetMapping("/challenge/{id}")
-    public HttpResponseApi<ChallengeDto> getChallenge(@PathVariable int id) {
+    public HttpResponseApi<ChallengeDto> getChallenge(@PathVariable Long id) {
         try {
             ChallengeDto challenge = challengeService.getById(id);
             return HttpResponseApi.<ChallengeDto>builder()
